Changing atlassian JIRA SDK to official atlassian-python-api SDK
