+++ airflow/providers/atlassian/jira/provider.yaml
@@ -38,18 +38,18 @@ integrations:
 operators:
   - integration-name: Atlassian Jira
     python-modules:
-      - airflow.providers.atlassian.operators.jira
+      - airflow.providers.atlassian.jira.operators.jira
 
 sensors:
   - integration-name: Atlassian Jira
     python-modules:
-      - airflow.providers.atlassian.sensors.jira
+      - airflow.providers.atlassian.jira.sensors.jira
 
 hooks:
   - integration-name: Atlassian Jira
     python-modules:
-      - airflow.providers.atlassian.hooks.jira
+      - airflow.providers.atlassian.jira.hooks.jira
 
 connection-types:
-  - hook-class-name: airflow.providers.atlassian.hooks.jira.JiraHook
+  - hook-class-name: airflow.providers.atlassian.jira.hooks.jira.JiraHook
     connection-type: jira
